düzenlemeler devam ediyor

+++ src/main/java/com/eren/aethra/daos/ModelDao.java
@@ -3,5 +3,5 @@ package com.eren.aethra.daos;
 import com.eren.aethra.models.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModelDao extends JpaRepository<Item, Integer> {
+public interface ModelDao extends JpaRepository<Item, Long> {
 }
